requirements.txt
# LangChain / LangGraph stack
langchain-core>=0.3
langgraph>=0.2
langchain-anthropic>=0.3
langchain-chroma>=0.1
langchain-text-splitters>=0.3
langchain-huggingface>=0.1
# langchain-aws>=0.2          # uncomment for Bedrock (ChatBedrockConverse)

# Config + logging
pydantic>=2.5
pydantic-settings>=2.1
structlog>=24.1
python-dotenv>=1.0

# Document parsing
pymupdf>=1.24
markdown-it-py>=3.0

# Embeddings (local default)
sentence-transformers>=3.0
# openai>=1.30                # uncomment if EMBEDDING_PROVIDER=openai
# boto3>=1.34                 # uncomment if any provider=bedrock

# Vector store backend (langchain-chroma pulls this transitively, listing for clarity)
chromadb>=0.5

# UI
streamlit>=1.36

# Dev
pytest>=8.0
ruff>=0.5
